add reason & int %

+++ eos_tax/views.py
@@ -12,5 +12,5 @@ from eos_tax.util import get_dates
 def index(request):
     dates = get_dates()
     website_data = get_website_data(dates=dates)
-    context = {"title":"Taxes to pay: " + str(TAX_RATE*100) + "%", "website_data":website_data }
+    context = {"title":"Taxes to pay: " + str(int(TAX_RATE*100)) + "%", "website_data":website_data }
     return render(request, "eos_tax/index.html", context)
